Split transformations core.py module into basic.py and composite.py
