+++ tests/test_targets.py
@@ -7,7 +7,7 @@ from pytest import param
 
 from baybe.exceptions import IncompatibilityError
 from baybe.targets.numerical import NumericalTarget
-from baybe.transformations.core import AffineTransformation
+from baybe.transformations.basic import AffineTransformation
 from baybe.utils.interval import Interval
